:art: fix https://github.com/siyuan-note/siyuan/issues/8979
