+++ app/appearance/langs/en_US.json
@@ -614,8 +614,7 @@
   "superBlock": "Super block",
   "embedBlock": "Embed block",
   "expand": "Expand",
-  "showDock": "Show Dock",
-  "hideDock": "Hide Dock",
+  "toggleDock": "Show/Hide Dock",
   "notebookName": "Please enter a notebook name",
   "moveToLeftTop": "Move to the Left Top",
   "moveToLeftBottom": "Move to the Left Bottom",
