+++ app/appearance/langs/es_ES.json
@@ -614,8 +614,7 @@
   "superBlock": "Superbloque",
   "embedBlock": "Bloque incrustado",
   "expand": "Expandir",
-  "showDock": "Mostrar Dock",
-  "hideDock": "Ocultar Dock",
+  "toggleDock": "Mostrar/ocultar Dock",
   "notebookName": "Por favor, introduzca un nombre de cuaderno",
   "moveToLeftTop": "Muéve a la parte superior izquierda",
   "moveToLeftBottom": "Muéve a la parte inferior izquierda",
