+++ app/appearance/langs/fr_FR.json
@@ -614,8 +614,7 @@
   "superBlock": "Superblock",
   "embedBlock": "Bloc intégré",
   "expand": "Élargir",
-  "showDock": "Montrer le Dock",
-  "hideDock": "Cacher le  Dock",
+  "toggleDock": "Afficher/masquer le Dock",
   "notebookName": "Veuillez entrer un nom de carnet de notes",
   "moveToLeftTop": "Déplacer vers la Gauche à Haut",
   "moveToLeftBottom": "Déplacer vers la Gauche à Bas",
