+++ app/appearance/langs/zh_CHT.json
@@ -614,8 +614,7 @@
   "superBlock": "超級塊",
   "embedBlock": "嵌入塊",
   "expand": "展開",
-  "showDock": "顯示停靠欄",
-  "hideDock": "隱藏停靠欄",
+  "toggleDock": "顯示/隱藏停靠欄",
   "notebookName": "請輸入筆記本名稱",
   "moveToLeftTop": "移動到左側上方",
   "moveToLeftBottom": "移動到左側下方",
