+++ app/appearance/langs/zh_CN.json
@@ -614,8 +614,7 @@
   "superBlock": "超级块",
   "embedBlock": "嵌入块",
   "expand": "展开",
-  "showDock": "显示停靠栏",
-  "hideDock": "隐藏停靠栏",
+  "toggleDock": "显示/隐藏停靠栏",
   "notebookName": "请输入笔记本名称",
   "moveToLeftTop": "移动到左侧上方",
   "moveToLeftBottom": "移动到左侧下方",
